My Food Ranger Website
FOOD  RANGER  is a premium food delivery (responsive website) with the mission to bring affordable and healthy meals to as many people as possible .This website is made by me with the help of courses and Google. This website mainly composed of html , CSS , JQUERY , media query etc. URL of WEBSITE- "(http://127.0.0.1:2902/index.html)"